now testing north and south hemisphere #233
